Type changes

+++ core/Clause.cpp
@@ -8,7 +8,7 @@
  *  i. ull n_literals is the number of literals in the clause
  * ii. ull* literals is an array of size n_literals where each element corresponds to a literal in the clause
  */
-Clause::Clause(ull* literals, ull n_literals){
+Clause::Clause(uint32_t* literals, int n_literals){
     this->literals = literals;
     this->n_literals = n_literals;
 }
@@ -17,7 +17,7 @@ Clause::Clause(ull* literals, ull n_literals){
  * instance) satisfies or not the SAT instance. Returns true if the clause is NOT satisfied, and returns false otherwise.
  */
 bool Clause::is_not_satisfied(VariablesArray* var_arr) const{
-    for(ull i = 0; i < n_literals; i++){ // For every literal in the clause...
+    for(int i = 0; i < n_literals; i++){ // For every literal in the clause...
         /* Let v = literals[i] >> 1 (i.e. v is the variable associated with the literal).
          * If the i^th literal is encoded to an odd number (literals[i] & 1) then check if !(var_arr->vars)[v] is true,
          * else check if (var_arr->vars)[v] is true.
